add save cookie file in sdk
